+++ shape/shape.go
@@ -1,5 +1,5 @@
 package shape
 
 type Shape interface {
-	Area() float32
+	Area() float64
 }
